+++ apps/manager/web/html/index.html
@@ -46,19 +46,19 @@
     <link rel="stylesheet" href="dispatcher.css" />
     <link rel="stylesheet" href="table.css" />
 
-    <script src="/grid.js"></script>
-    <script src="/table.js"></script>
-    <script src="/webSocketHandler.js"></script>
-    <script src="/view.js"></script>
+    <script src="/core/grid.js"></script>
+    <script src="/core/table.js"></script>
+    <script src="/core/webSocketHandler.js"></script>
+    <script src="/core/view.js"></script>
     <script src="/views/dispatcher.js"></script>
-    <script src="/helper.js"></script>
+    <script src="/help/helper.js"></script>
     <script src="/views/metadataCollector.js"></script>
-    <script src="/duration.js"></script>
-    <script src="/timer.js"></script>
-    <script src="/select.js"></script>
+    <script src="/core/duration.js"></script>
+    <script src="/core/timer.js"></script>
+    <script src="/core/select.js"></script>
     <script src="/views/appView.js"></script>
     <script src="/views/createdApps.js"></script>
-    <script src="/dynamic.js"></script>
+    <script src="/help/dynamic.js"></script>
     <script src="/views/account.js"></script>
     <script src="/views/gridSettingsView.js"></script>
     <script src="/views/installedApps.js"></script>
