+++ apps/manager/web/js/core/duration.js
