+++ apps/manager/web/js/core/timer.js
