+++ apps/manager/web/js/help/dynamic.js
