+++ apps/manager/web/js/help/helper.js
