+++ apps/manager/web/js/views/metadataCollector.js
@@ -20,7 +20,7 @@
  */
 
 if (typeof module !== "undefined" && typeof module.exports !== "undefined") {
-	View = require("../view");
+	View = require("../core/view");
 	Dispatcher = require("./dispatcher").Dispatcher;
 }
